+++ app/components/account/details/programatic-usage/aad-credentials-picker/aad-app-picker/index.ts
@@ -1,0 +1,1 @@
+export * from "./aad-app-picker.component";
