+++ app/components/account/details/programatic-usage/aad-credentials-picker/aad-credentials-picker.html
@@ -1,29 +1,8 @@
-<div class="app-lists">
-    <div class="heading">
-        <h2>Step 1: Pick the application you want to use</h2>
+<ng-container [ngSwitch]="currentStep">
+    <bl-aad-app-picker *ngSwitchCase="Step.pickApplication" (selectedApplicationChanged)="pickApplication($event)"></bl-aad-app-picker>
+    <div *ngSwitchCase="Step.generateSecret">
+        <ng-container *ngIf="pickedApplication">
+            Gen secrets for {{pickedApplication.displayName}}
+        </ng-container>
     </div>
-    <div class="create-new">
-        <bl-button type="wide" color="primary" >Create new</bl-button>
-    </div>
-    <div class="search">
-        <input [formControl]="searchAppControl" placeholder="Search">
-    </div>
-    <div class="app-list-title">
-        With permissions
-    </div>
-    <div class="app-list-title">
-        All apps
-    </div>
-    <bl-table (activeItemChange)="pickApplication($event)" class="app-list apps-with-permission">
-        <bl-row *ngFor="let app of apps; trackBy: trackApp" [key]="app.id">
-            <bl-cell [value]="app.displayName"></bl-cell>
-            <bl-cell [value]="getPermission(app.id)" style="width: 40px"></bl-cell>
-        </bl-row>
-    </bl-table>
-
-    <bl-table (activeItemChange)="pickApplication($event)" class="app-list apps-with-permission">
-        <bl-row *ngFor="let app of allApps; trackBy: trackApp" [key]="app.id">
-            <bl-cell [value]="app.displayName"></bl-cell>
-        </bl-row>
-    </bl-table>
-</div>
+</ng-container>
