+++ app/components/account/details/programatic-usage/aad-credentials-picker/generate-aad-app-secret/generate-aad-app-secret.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from "@angular/core";
+
+import { ServicePrincipal } from "app/models/ms-graph";
+import "./generate-aad-app-secret.scss";
+
+@Component({
+    selector: "bl-generate-aad-app-secret",
+    templateUrl: "generate-aad-app-secret.html",
+})
+export class GenerateAADAppSecretComponent {
+    @Input() public app: ServicePrincipal;
+}
