+++ app/components/account/details/programatic-usage/aad-credentials-picker/generate-aad-app-secret/index.ts
@@ -1,0 +1,1 @@
+export * from "./generate-aad-app-secret.component";
