+++ app/components/account/details/programatic-usage/aad-credentials-picker/index.ts
@@ -1,1 +1,3 @@
+export * from "./generate-aad-app-secret";
+export * from "./aad-app-picker";
 export * from "./aad-credentials-picker.component";
