+++ app/components/account/details/programatic-usage/index.ts
@@ -1,2 +1,3 @@
 export * from "./programatic-usage.component";
 export * from "./aad-credentials-picker";
+export * from "./programatic-usage.module";
