+++ app/components/account/details/programatic-usage/programatic-usage.module.ts
@@ -1,0 +1,38 @@
+import { NgModule } from "@angular/core";
+import { commonModules } from "app/common";
+import {
+    AADAppPickerComponent, AADCredentialsPickerComponent, GenerateAADAppSecretComponent,
+} from "./aad-credentials-picker";
+import { ProgramaticUsageComponent } from "./programatic-usage.component";
+
+const privateComponents = [
+    AADCredentialsPickerComponent,
+    AADAppPickerComponent,
+    GenerateAADAppSecretComponent,
+];
+console.log("PRivate", privateComponents);
+
+const publicComponents = [
+    ProgramaticUsageComponent,
+];
+
+const modules = [
+    ...commonModules,
+];
+
+@NgModule({
+    declarations: [
+        ...privateComponents,
+        ...publicComponents,
+    ],
+    exports: [
+        ...publicComponents,
+    ],
+    entryComponents: [
+        ProgramaticUsageComponent,
+    ],
+    imports: modules,
+})
+export class ProgramaticUsageModule {
+
+}
